Move agent logic from rpc/agent to cmd/plakar/subcommands/agent
